Move Auth and Microservcies module to separate packages
