+++ .github/workflows/passport.test.yaml
@@ -49,7 +49,7 @@ jobs:
       - uses: pnpm/action-setup@v2
         with:
           version: 8
-      - run: pnpm -F @looport/passport install
+      - run: pnpm -F @looport/passport install && pnpm -F @looport/passport^... build
       - run: pnpm -F @looport/passport lint
   type-check:
     runs-on: ubuntu-22.04
@@ -61,7 +61,7 @@ jobs:
       - uses: pnpm/action-setup@v2
         with:
           version: 8
-      - run: pnpm -F @looport/passport install
+      - run: pnpm -F @looport/passport install && pnpm -F @looport/passport^... build
       - run: pnpm -F @looport/passport type-check
   build:
     runs-on: ubuntu-22.04
@@ -73,5 +73,4 @@ jobs:
       - uses: pnpm/action-setup@v2
         with:
           version: 8
-      - run: pnpm -F @looport/passport install
-      - run: pnpm -F @looport/passport build
+      - run: pnpm -F @looport/passport install && pnpm -F @looport/passport... build
