+++ application/services/passport/src/app/services/auth-config/auth-config.service.ts
@@ -3,7 +3,7 @@ import {Injectable} from '@nestjs/common'
 import {ConfigService} from '@nestjs/config'
 
 @Injectable()
-export class AuthenticationConfigService {
+export class AuthConfigService {
   constructor(private readonly configService: ConfigService) {}
 
   getAuthConfig(): AuthOptions {
