+++ application/services/telegraph/src/app/services/auth-config/auth-config.service.ts
@@ -1,0 +1,17 @@
+import {AuthOptions} from '@looport/nest-auth'
+import {Injectable} from '@nestjs/common'
+import {ConfigService} from '@nestjs/config'
+
+@Injectable()
+export class AuthConfigService {
+  constructor(private readonly configService: ConfigService) {}
+
+  getAuthConfig(): AuthOptions {
+    return {
+      jwt: {
+        secret: this.configService.getOrThrow('JWT_SECRET'),
+        signOptions: {expiresIn: this.configService.getOrThrow('JWT_EXPIRES')},
+      },
+    }
+  }
+}
