+++ package.json
@@ -14,6 +14,7 @@
     "web": "pnpm --filter '@looport/web' dev",
     "passport": "pnpm --filter '@looport/passport' start:dev",
     "messenger": "pnpm --filter '@looport/messenger' start:dev",
-    "format": "pnpm --filter '*' format"
+    "format": "pnpm --filter '*' format",
+    "test:all": "pnpm --filter '*' test:all"
   }
 }
